# app-package/openapi/browser-tools.yml
openapi: 3.0.3
info:
  title: Secure Browser Tools API
  version: 1.0.0
servers:
  - url: https://agent.example.com
paths:
...
  /api/skills/orchestrate_workflow:
    post:
      operationId: orchestrate_workflow
      summary: Execute a multi-step workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK